Add DB connection, skills controller working. Start Front
