+++ nine-gates/server/index.js
@@ -1,14 +1,18 @@
+
 const express = require('express');
-const searchRoute = require('./routes/search');
+const mongoose = require('mongoose');
+const jobController = require('./controllers/jobController');
+
+// Connect to the MongoDB database
+mongoose.connect('mongodb://localhost:27017/jobSkillsDB')
+  .then(() => console.log('Database connected successfully'))
+  .catch(err => console.error('Database connection error: ', err));
+
 const app = express();
-const port = 5000;
 
-app.use('/search', searchRoute);
+// Set up your routes
+app.get('/search/:job', jobController.getJobSkills);
 
-app.get('/', (req, res) => {
-  res.send('Hello World!');
-});
-
-app.listen(port, () => {
-  console.log(`Server is running on http://localhost:${port}`);
-});
+// Start the server
+const port = process.env.PORT || 5000;
+app.listen(port, () => console.log(`Server is running on port ${port}`));
