+++ nine-gates/server/models/Job.js
@@ -1,0 +1,8 @@
+const mongoose = require('mongoose');
+
+const jobSchema = new mongoose.Schema({
+  jobTitle: String,
+  essentialSkills: [String]
+});
+
+module.exports = mongoose.model('Job', jobSchema);
