+++ nine-gates/server/package.json
@@ -1,7 +1,9 @@
 {
   "dependencies": {
     "axios": "^1.7.2",
+    "cors": "^2.8.5",
     "express": "^4.19.2",
+    "mongoose": "^8.4.1",
     "nodemon": "^3.1.3"
   }
 }
